requirements.txt
flake8==3.5.0
gitdb2==2.0.3
GitPython==2.1.10
ipgetter==0.7
mccabe==0.6.1
paho-mqtt==1.3.1
pycodestyle==2.3.1
pycryptodome==3.6.1
pyflakes==1.6.0
smmap2==2.0.3
